backend/requirements.txt
fastapi
uvicorn[standard]
spacy
pymupdf
requests
beautifulsoup4
sqlalchemy
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
python-multipart
